+++ packages/outbox/src/repositories/jobs-outbox.repository.ts
@@ -3,15 +3,15 @@ import {
   JobsOutboxEntity,
   JobsOutboxModel,
   type JobType,
-  type JobPayload,
   type Repository,
 } from '@volontariapp/database';
 import type { DataSource } from 'typeorm';
 
-export class JobsOutboxRepository<
-  K extends JobType = JobType,
-  P = JobPayload<K>,
-> extends BaseRepository<JobsOutboxModel, JobsOutboxEntity<K, P>, string> {
+export class JobsOutboxRepository<K extends JobType = JobType> extends BaseRepository<
+  JobsOutboxModel,
+  JobsOutboxEntity<K>,
+  string
+> {
   constructor(dataSourceOrRepository: DataSource | Repository<JobsOutboxModel>) {
     if ('getRepository' in dataSourceOrRepository) {
       super(