+++ packages/outbox/src/test/utils/repositories/jobs-outbox-test.repository.ts
@@ -3,14 +3,14 @@ import {
   JobsOutboxEntity,
   JobsOutboxModel,
   type JobType,
-  type JobPayload,
   type Repository,
 } from '@volontariapp/database';
 
-export class TestJobsOutboxRepository<
-  K extends JobType = JobType,
-  P = JobPayload<K>,
-> extends BaseRepository<JobsOutboxModel, JobsOutboxEntity<K, P>, string> {
+export class TestJobsOutboxRepository<K extends JobType = JobType> extends BaseRepository<
+  JobsOutboxModel,
+  JobsOutboxEntity<K>,
+  string
+> {
   constructor(repository: Repository<JobsOutboxModel>) {
     super(repository, JobsOutboxEntity, JobsOutboxModel);
   }